Humans and nature, nature and fashion - Mymoderno
Our future is in a very strong connection between humans and nature. The trials facing humanity today (Covid 19, natural disasters, pollution ...) show us the only way - the balance between nature and us. Choosing quality, not quantity.
The connection between nature and fashion can also be very strong. Natural materials and handmade materials in the creation of clothing and shoes increase their quality and prolong their use. This in turn leads to a reduction in overproduction and consumption, and then to a cleaner environment.
The environmentally friendly way of life leads us to appreciate natural products, quality natural cosmetics and eco and organic foods. To be healthy, our mother Earth must be healthy - to love, cherish and cherish her, so that we can enjoy her beauty for a long time. And the unique gifts of nature to be our muse to create and create beautiful embroidered jewelry and accessories - necklaces, bracelets, buttons, neckties; unique handmade toys - giraffe, bee yeast, dog, butterfly girl; carvings - horses, carved icon Jesus Christ, beautiful woman, rosettes for ceiling; handmade wooden chess and wicker antique baskets; handmade designer shoes and slippers, high-quality sports and classic shirts, hand-knitted bags, gloves and hats, purses, keychains and tablet cases, handmade. Original branded clothing, shoes and accessories.

Handmade shoes and boots feature exceptional quality and durability, as well as wallets, key chains and genuine leather tablet cases. The unique properties of natural lambskin - "breathing" - regulating body temperature no matter the season - producing heat in winter and keeping the body cool in the summer are cleverly combined in chairs for lambskin chairs and car mats, as well as in home slippers and boots made of sheepskin. Knitted antique baskets are the perfect home accessory, also suitable for a gift, a cat or a small pet.
Wood carving - with the softness and warmth of the wood, it brings coziness and classic look to any home. A carved icon or rosette for a ceiling, a wooden horse or a beautiful woman, wood carving is the most beautiful gift that reminds of the power and beauty of the forest, as well as our duty to protect and cherish nature.
Choosing quality, not quantity. To love, appreciate and cherish nature so that we can enjoy its beauty for a long time. So the beautiful natural elements and shapes that inspire craftsmen to make the products will be preserved and they will continue to create. And free shipping.